🛠️: refactor `FormOptionalText` test to typescript
